+++ tests/Acme81/Attribute/SampleNestedValue.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Acme81\Attribute;
+
+class SampleNestedValue
+{
+    public function __construct(
+        public int $value
+    ) {
+    }
+}