+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # CHANGELOG
 
+## v2.2.0
+
+### New Requirements
+
+None
+
+### New features
+
+None
+
+### Deprecated Features
+
+None
+
+### Backward Incompatible Changes
+
+None
+
+### Other Changes
+
+- [#36](https://github.com/olvlvl/composer-attribute-collector/pull/36) Attribute arguments are now serialized to support [nested attributes introduced in PHP 8.1](https://wiki.php.net/rfc/new_in_initializers).
+
+
+
 ## v2.1.0
 
 ### New Requirements